Move WASM API into single file
